src/main/java/com/my/project/petclinic/hospital/domain/service/PatientService.java
package com.my.project.petclinic.hospital.domain.service;

import com.my.project.petclinic.hospital.domain.model.Patient;
import com.my.project.petclinic.hospital.persistence.PatientRepository;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@AllArgsConstructor
@Service
public class PatientService {

    private final PatientRepository repository;

    public List<Patient> getAllPatients() {
        return repository.findAll();
    }

    public Patient save(Patient patient) {
        return repository.save(patient);
    }

    public Patient update(Patient patient) {
        return repository.update(patient);
    }
}
